code/intel-station/requirements.txt
streamlit==1.56.0
python-dotenv==1.1.0
pyyaml==6.0.2
